user/serializers.py
from rest_framework import serializers
from user.models import User


class UserMinimalSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ("id", "username", "role")

